Move HostRuntime to be WorkerHarness.
